Refactored things so that IMW::Files::File objects no longer include the IMW::Files::Compresible module.  This was leading to a circular set of `require' statements.  File objects can no longer be compressed, but the gain is that the circularity is avoided without having to resort to `load' statements and CompressedFile objects don't have compress methods, which is nice.  Each subclass of File will have to itself include the Compressible module.  This is probably fine as two (three?) subclasses immediately suggest themselves: `Text' and `Binary' (perhaps `Unicode'?) and all other file types (csv, xml, &c.) can subclass these, so any common file will still wind up with compression methods if only the `Text' and `Binary' classes mixin the Compressible module.

+++ lib/imw/model/files/compressed_file.rb
@@ -20,7 +20,7 @@
 # <tt>:decompression_flags</tt>:: a string of flags to pass to the
 # compression program when decompressing the file.
 #
-# A subclass must also define the method +decompressed_file+ which
+# A subclass must also define the method +decompressed_path+ which
 # returns the path of the file post-decompression.
 #
 # Author::    (Philip flip Kromer, Dhruv Bansal) for Infinite Monkeywrench Project (mailto:coders@infochimps.org)
@@ -29,7 +29,10 @@
 # Website::   http://infinitemonkeywrench.org/
 # 
 
+require 'fileutils'
+
 require 'imw/model/files/file'
+require 'imw/model/files/identify'
 
 require 'imw/utils'
 
@@ -48,13 +51,14 @@ module IMW
         [IMW::EXTERNAL_PROGRAMS[@compression[:program]],@compression[:decompression_flags],@path].join ' '
       end
         
-      protected
+      public
       # Decompress this file in its present directory overwriting any
       # existing files and without saving the original compressed
       # file.
       def decompress!
-        IMW.system decompression_command
-        decompressed_file
+        raise IMW::PathError.new("cannot decompress #{@path}, doesn't exist!") unless exist?
+        FileUtils.cd(@dirname) { IMW.system decompression_command }
+        IMW::Files.identify(decompressed_path)
       end
 
       # Decompress this file in its present directory, overwriting any
@@ -63,10 +67,14 @@ module IMW
       # The implementation is a little stupid, as the file is
       # needlessly copied.
       def decompress
-        FileUtils.cp(@path,@path + 'copy')
-        decompress!
-        FileUtils.mv(@path + 'copy',@path)
-        decompressed_file
+        raise IMW::PathError.new("cannot decompress #{@path}, doesn't exist!") unless exist?        
+        begin
+          FileUtils.cp(@path,@path + 'copy')
+          decompress!
+        ensure
+          FileUtils.mv(@path + 'copy',@path)
+        end
+        IMW::Files.identify(decompressed_path)
       end
 
     end
